requirements.txt
django >3.2
python-decouple
pillow
djangorestframework
markdown
django-filter